+++ camera-qr/src/wasmJsMain/kotlin/tech/kotlinlang/camera/qr/QrCodeImageAnalyserHolder.wasmJs.kt
@@ -1,0 +1,10 @@
+package tech.kotlinlang.camera.qr
+
+import tech.kotlinlang.camera.analyser.ImageAnalyser
+
+actual fun getQrCodeImageAnalyser(): ImageAnalyser<String> {
+    return object: ImageAnalyser<String> {
+        override fun setListener(callback: (String) -> Unit) {
+        }
+    }
+}