+++ camera/src/wasmJsMain/kotlin/tech/kotlinlang/camera/Util.wasmJs.kt
@@ -1,0 +1,5 @@
+package tech.kotlinlang.camera
+
+actual fun getCameraHelper(): CameraHelper {
+    return WasmCameraHelper()
+}