+++ camera/src/wasmJsMain/kotlin/tech/kotlinlang/camera/analyser/EmptyImageAnalyserHolder.wasmJs.kt
@@ -1,0 +1,9 @@
+package tech.kotlinlang.camera.analyser
+
+actual fun getEmptyImageAnalyser(): ImageAnalyser<Unit> {
+    return object: ImageAnalyser<Unit> {
+        override fun setListener(callback: (Unit) -> Unit) {
+
+        }
+    }
+}